fix: validate fsync fd argument; advance wb_err snapshot on error paths
CodeRabbit Major findings on PR #775:

1. The dispatch arms for SYS_FSYNC / SYS_FDATASYNC silently truncated
   the raw u64 arg to u32, so an fd like (1u64 << 32) | 3 would alias
   to fd 3. Validate `fd_raw > u32::MAX` up front and return EBADF.

2. `OpenFile::do_fsync` only advanced the per-file wb_err snapshot on
   the success path. If the FileOps::fsync hook or SuperOps::sync_fs
   returned an error, the snapshot stayed stale and a *subsequent*
   fsync could re-surface a since-consumed writeback EIO. Fold both
   error paths through a `consume_wb_err` closure so every exit
   advances the snapshot exactly once.

+++ kernel/src/arch/x86_64/syscall.rs
@@ -608,13 +608,16 @@ pub unsafe extern "C" fn syscall_dispatch(
 
         // fsync(fd) — flush page cache + per-mount BlockCache::sync_fs;
         // surface sticky EIO via the per-OpenFile errseq snapshot. RFC
-        // 0007 §Ordering vs fsync/fdatasync.
-        FSYNC => super::syscalls::vfs::sys_fsync_impl(a0 as u32, false),
+        // 0007 §Ordering vs fsync/fdatasync. The raw `a0` (u64) is
+        // passed through; the impl validates the high 32 bits to
+        // reject e.g. `fsync(0x1_0000_0003)` rather than silently
+        // truncating to fd=3.
+        FSYNC => super::syscalls::vfs::sys_fsync_impl(a0, false),
 
         // fdatasync(fd) — same data flush as fsync, may skip the
         // inode-table flush per Linux semantics. See sys_fsync_impl
         // for the data_only-vs-fsync split.
-        FDATASYNC => super::syscalls::vfs::sys_fsync_impl(a0 as u32, true),
+        FDATASYNC => super::syscalls::vfs::sys_fsync_impl(a0, true),
 
         // dup2(oldfd, newfd)
         DUP2 => {
